move vector comparison extension into separate module
